@@ -1,18 +1,21 @@
 import ComposableArchitecture
 import DataClient
-import XCTest
+import SharedModels
+import Testing
 
 @testable import ScheduleFeature
 
-final class ScheduleTests: XCTestCase {
-  @MainActor
-  func testFetchData() async {
+@Suite
+@MainActor
+struct ScheduleTests {
+  @Test
+  func fetchData_success() async {
     let store = TestStore(initialState: Schedule.State()) {
       Schedule()
     } withDependencies: {
-      $0[DataClient.self].fetchDay1 = { @Sendable in .mock1 }
-      $0[DataClient.self].fetchDay2 = { @Sendable in .mock2 }
-      $0[DataClient.self].fetchDay3 = { @Sendable in .mock3 }
+      $0[DataClient.self].fetchDay1 = { @Sendable _ in .mock1 }
+      $0[DataClient.self].fetchDay2 = { @Sendable _ in .mock2 }
+      $0[DataClient.self].fetchDay3 = { @Sendable _ in .mock3 }
     }
     await store.send(.view(.onAppear))
     await store.receive(\.fetchResponse.success) {
@@ -22,15 +25,15 @@ final class ScheduleTests: XCTestCase {
     }
   }
 
-  @MainActor
-  func testFetchDataFailure() async {
+  @Test
+  func fetchData_failure() async {
     struct FetchError: Equatable, Error {}
     let store = TestStore(initialState: Schedule.State()) {
       Schedule()
     } withDependencies: {
-      $0[DataClient.self].fetchDay1 = { @Sendable in throw FetchError() }
-      $0[DataClient.self].fetchDay2 = { @Sendable in .mock2 }
-      $0[DataClient.self].fetchDay3 = { @Sendable in .mock3 }
+      $0[DataClient.self].fetchDay1 = { @Sendable _ in throw FetchError() }
+      $0[DataClient.self].fetchDay2 = { @Sendable _ in .mock2 }
+      $0[DataClient.self].fetchDay3 = { @Sendable _ in .mock3 }
     }
     await store.send(.view(.onAppear))
     await store.receive(\.fetchResponse.failure)
